requirements.txt
Flask==3.0.2
pillow==10.2.0